refactor(backend): move ProviderSkillCall and isProviderSkillMessage to util
Fixes layer violation where a util module imported from a service module.
The canonical definitions now live in util/provider-skill.ts, with
re-exports from llm-client.ts for backwards compatibility.
